# openapi.yaml
openapi: 3.1.0
...
info:
  title: LAGALia — Gestió de contractació pública
  version: 0.1.0
  description: |
    API de la plataforma de gestió de contractació pública de l'Ajuntament de Cunit.

    **API first**: tota funcionalitat del sistema és accessible per aquesta API.
    La interfície web n'és un client més, igual que els agents d'IA, n8n o
    qualsevol integració de tercers.

    Convencions generals a `docs/05-api.md`.
  contact:
    name: Ajuntament de Cunit
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
servers:
  - url: /api/v1
    description: Servidor actual
...
security:
  - bearerAuth: []
paths:
...
  /contracts:
    get:
      tags: [contracts]
      operationId: listContracts
      summary: Llistat de contractes (abast departamental)
      description: |
        L'abast s'aplica al servidor: amb accés departamental només es veuen
        els contractes dels departaments propis o dels quals s'és responsable.
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/ViewScope"
        - name: q
          in: query
          description: Cerca lliure per expedient, objecte o adjudicatari.
          schema: { type: string, maxLength: 200 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [published_at, -published_at, calculated_end_date, -calculated_end_date,
                   award_amount, -award_amount, file_code, -file_code]
            default: "-published_at"
        - name: filter[department_id]
          in: query
          schema: { type: integer, format: int64 }
        - name: filter[unassigned]
          in: query
          description: Només contractes sense cap departament assignat.
          schema: { type: boolean }
        - name: filter[contract_type]
          in: query
          schema: { type: string }
        - name: filter[status]
          in: query
          schema: { type: string }
        - name: filter[internal_status]
          in: query
          schema: { type: string, enum: [normal, pending_review, approved, rejected] }
        - name: filter[expiry_warning]
          in: query
          schema: { type: boolean }
        - name: filter[possibly_finished]
          in: query
          schema: { type: boolean }
        - name: filter[year]
          in: query
          description: Exercici (any de publicació).
          schema: { type: integer }
        - name: filter[contractor_id]
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Pàgina de contractes
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ContractSummary" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [contracts]
      operationId: createContract
      summary: Alta manual d'un contracte
      x-required-scopes: [contracts:create]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContractCreate" }
      responses:
        "201":
          description: Contracte creat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Ja existeix un contracte amb la mateixa clau natural
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /contracts/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [contracts]
      operationId: getContract
      summary: Detall d'un contracte (amb lots germans i comptadors)
      description: Fora de l'abast departamental la resposta és `404`.
      x-required-scopes: [contracts:read]
      responses:
        "200":
          description: Contracte
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [contracts]
      operationId: updateContract
      summary: Edició parcial segons la matriu de permisos
      description: |
        Amb `contracts:update` (admin, gestió de contractació) tots els camps;
        amb només `contracts:update_warning` (responsable de departament, dins
        del seu abast), únicament `warning_months_override`.
      x-required-scopes: [contracts:update, contracts:update_warning]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContractUpdate" }
      responses:
        "200":
          description: Contracte actualitzat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /contracts/{id}/history:
    get:
      tags: [contracts]
      operationId: getContractHistory
      summary: Historial de canvis
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Historial (descendent)
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ContractHistoryEntry" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/{id}/extensions:
    get:
      tags: [contracts]
      operationId: getContractExtensions
      summary: Pròrrogues del contracte
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Pròrrogues
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ContractExtension" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/{id}/modifications:
    get:
      tags: [contracts]
      operationId: getContractModifications
      summary: Modificacions del contracte
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Modificacions
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ContractModification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/stats:
    get:
      tags: [contracts]
      operationId: getContractsStats
      summary: KPIs del dashboard (dins l'abast de l'usuari)
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ViewScope"
        - name: filter[year]
          in: query
          schema: { type: integer }
        - name: filter[amount_min]
          in: query
          schema: { type: number, minimum: 0 }
        - name: filter[amount_max]
          in: query
          schema: { type: number, minimum: 0 }
      responses:
        "200":
          description: Estadístiques
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContractStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /contracts/facets:
    get:
      tags: [contracts]
      operationId: getContractsFacets
      summary: Valors distincts per als filtres del llistat
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ViewScope"
      responses:
        "200":
          description: Facets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContractFacets" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /contracts/{id}/actions/finish:
    post:
      tags: [contracts]
      operationId: finishContract
      summary: Finalitza el contracte (estat → Finalitzat)
      x-required-scopes: [contracts:close_alert]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Contracte actualitzat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /contracts/{id}/actions/dismiss-expiry:
    post:
      tags: [contracts]
      operationId: dismissContractExpiry
      summary: Descarta l'alerta de venciment (persistent fins que canviï la data final)
      x-required-scopes: [contracts:close_alert]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Contracte actualitzat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /contracts/{id}/actions/enrich:
    post:
      tags: [contracts]
      operationId: enrichContract
      summary: Encua l'enriquiment del contracte (job enrich.contract)
      x-required-scopes: [contracts:enrich]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "202":
          description: Job encuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /contracts/exports:
    post:
      tags: [contracts]
      operationId: createContractsExport
      summary: Encua l'exportació del llistat filtrat (CSV o XLSX)
      description: |
        L'abast efectiu (view + rol) es resol a l'encuament: l'export conté
        només el que l'usuari veu. La descàrrega es fa amb token efímer.
      x-required-scopes: [contracts:export]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExportRequest" }
      responses:
        "202":
          description: Job d'exportació encuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /contracts/exports/{id}/download:
    get:
      tags: [contracts]
      operationId: downloadContractsExport
      summary: Descarrega el fitxer d'una exportació completada
      description: |
        Autorització per token efímer d'un sol ús (`POST /auth/ephemeral`,
        propòsit `download`, recurs = id del job). Mai el JWT de sessió.
      security: []
      parameters:
        - $ref: "#/components/parameters/ResourceUuid"
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      responses:
        "200":
          description: Fitxer exportat
          content:
            text/csv:
              schema: { type: string }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /contracts/bulk/assign-departments:
    post:
      tags: [contracts]
      operationId: bulkAssignDepartments
      summary: Assignació massiva de departaments
      x-required-scopes: [contracts:bulk_assign]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkAssignRequest" }
      responses:
        "200":
          description: Resum de l'assignació
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkAssignResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /contracts/{id}/criteria:
    get:
      tags: [contracts]
      operationId: getContractCriteria
      summary: Criteris d'adjudicació del contracte
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Criteris ordenats per posició
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/AwardCriterion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/{id}/committee:
    get:
      tags: [contracts]
      operationId: getContractCommittee
      summary: Mesa de contractació del contracte
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Membres de la mesa
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/CommitteeMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/{id}/documents:
    get:
      tags: [contracts]
      operationId: getContractDocuments
      summary: Documents de fase del contracte
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Documents ordenats per fase
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PhaseDocument" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /contracts/{id}/executions:
    get:
      tags: [contracts]
      operationId: getContractExecutions
      summary: Actuacions de la fase d'execució de l'expedient (dataset 8idu-wkjv)
      x-required-scopes: [contracts:read]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Actuacions, més recents primer
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ContractExecution" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ───────────────────────── minor-contracts ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token d'accés d'usuari (JWT curt) o clau de compte de servei (`sk_...`)
        per a agents d'IA i integracions. Sempre a la capçalera `Authorization`;
        mai per query string.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    PageSize:
      name: page[size]
      in: query
      description: Elements per pàgina.
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    PageCursor:
      name: page[cursor]
      in: query
      description: Cursor de la pàgina següent, retornat a `meta.next_cursor`.
      schema: { type: string }
    ViewScope:
      name: view
      in: query
      description: |
        Abast demanat. `user` (per defecte) limita als departaments propis;
        `all` demana abast complet i **només s'accepta** si el rol de l'usuari ho
        permet — es valida al servidor, no és una preferència del client.
      schema: { type: string, enum: [user, all], default: user }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Clau única del client per evitar duplicats en reintents. Repetir la
        mateixa clau retorna el resultat original en comptes de crear un recurs nou.
      schema: { type: string, maxLength: 128 }
  responses:
...
    Unauthorized:
      description: Cal autenticar-se o la sessió ha caducat
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Autenticat però sense permís per a aquesta acció o recurs
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationError:
      description: Dades d'entrada no vàlides
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    PageMeta:
      type: object
      required: [total]
      properties:
        total: { type: integer }
        next_cursor:
          type: [string, "null"]
...
    ContractSummary:
      type: object
      required: [id, file_code, status, lot, internal_status, expiry_warning, possibly_finished]
      properties:
        id: { type: integer, format: int64 }
        file_code: { type: string }
        status: { type: string }
        lot: { type: string }
        subject: { type: [string, "null"] }
        contract_type: { type: [string, "null"] }
        procedure: { type: [string, "null"] }
        contractor: { oneOf: [{ $ref: "#/components/schemas/ContractorRef" }, { type: "null" }] }
        award_amount: { type: [string, "null"], description: Decimal serialitzat }
        published_at: { type: [string, "null"], format: date-time }
        start_date: { type: [string, "null"], format: date }
        calculated_end_date: { type: [string, "null"], format: date }
        expiry_warning: { type: boolean }
        possibly_finished: { type: boolean }
        internal_status: { type: string, enum: [normal, pending_review, approved, rejected] }
        department_ids:
          type: array
          items: { type: integer, format: int64 }
...